+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Adminer dev
+- Do not drop the old check constraint when the new one cannot be created
 - MySQL: Display the check constraint clause without the extra escaping added by information_schema
 ### Plugins
 - New method assetUrl() returning the URL of a static file in the compiled Adminer